requirements.txt
# Security and Performance Enhancement Dependencies

# Core Machine Learning and Data Processing
numpy==1.21.5
pandas==1.3.5
scikit-learn==1.0.2
scipy==1.7.3

# Deep Learning and AI
tensorflow==2.8.0
keras==2.8.0
torch==1.10.2

# Cryptography and Security
cryptography==3.4.8
python-jose==3.3.0

# Database and ORM
sqlalchemy==1.4.36
psycopg2-binary==2.9.3

# Performance and Optimization
numba==0.55.1
joblib==1.1.0

# Encryption and Hashing
pycryptodome==3.14.1

# Logging and Monitoring
python-json-logger==2.0.2

# Testing and Validation
pytest==7.1.1
pytest-cov==3.0.0

# Type Checking
mypy==0.950
typing-extensions==4.2.0

# Development and Debugging
ipython==8.2.0
ipdb==0.13.9

# Caching and Performance
redis==4.3.4
hiredis==2.1.0  # Optional but recommended for better Redis performance

# Performance Testing
aiohttp==3.8.4
matplotlib==3.5.1
asyncio==3.4.3

# OverWatch TOSS implementation
plotly==5.7.0
websockets==10.2
h5py==3.6.0
Pillow==9.0.1
python-dotenv==0.19.2
PyYAML==6.0
seaborn==0.11.2

# Geospatial dependencies
geopandas==0.10.2
shapely==1.8.2
rasterio==1.2.10
pyproj==3.3.0
networkx==2.7.1

# Mobile App dependencies
kivy==2.1.0
plyer==2.0.0
cryptography==3.4.8
requests==2.27.1
sounddevice==0.4.4
pyobjus==1.2.1  # iOS support
jnius==1.1.0    # Android support
aiohttp==3.8.4

# Accounting and Financial System dependencies
sqlalchemy==1.4.36
psycopg2-binary==2.9.3
pytesseract==0.3.9
stripe==3.0.0
quickbooks==0.7.0
python-dateutil==2.8.2

# Security Hardening dependencies
bcrypt==4.0.1
cryptography==3.4.8
pyOpenSSL==22.0.0
certifi==2022.5.18.1
idna==3.3
requests==2.27.1
urllib3==1.26.9
pyjwt==2.4.0
python-iptools==0.7.0